+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/NPC/EffectSpawnGrieferAlien.lua
@@ -51,6 +51,7 @@ function EffectSpawnGrieferAlien:OnStart()
     end
 
     zombie:resetModelNextFrame()
+    npc:EnterPlayerVehicle(player)
 end
 
 function EffectSpawnGrieferAlien:OnEnd()