+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/NPC/EffectSpawnGrieferSanta.lua
@@ -33,4 +33,5 @@ function EffectSpawnGrieferSanta:OnStart()
     npc.npcGroup = ChaosNPCGroupID.RAIDERS
 
     npc:SetWeapon("Base.Sledgehammer")
+    npc:EnterPlayerVehicle(player)
 end