+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/NPC/EffectSpawnNPC.lua
@@ -35,4 +35,5 @@ function EffectSpawnNPC:OnStart()
 
     -- npc:SetWeapon("Base.PipeWrench")
     npc:SetWeapon("Base.BaseballBat")
+    npc:EnterPlayerVehicle(player)
 end