+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/Effects/List/NPC/EffectWomanizer.lua
@@ -23,6 +23,7 @@ function EffectWomanizer:OnStart()
                 zombie:dressInRandomOutfit()
                 npc:initializeHuman()
                 npc.npcGroup = ChaosNPCGroupID.RAIDERS
+                npc:EnterPlayerVehicle(player)
             end
         end
     end